requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sentence-transformers==3.1.1
faiss-cpu==1.9.0
numpy==1.26.4
scikit-learn==1.5.2
rank-bm25==0.2.2
PyMuPDF==1.24.8
python-dotenv==1.0.1
